Revert "[EASY] Move Scorer to analysis tools (#1364)"
This reverts commit b7ef5cf753143d5a1891241880aba27b7a4ce8c1.
