+++ docs/packages/classification.rst
@@ -18,6 +18,7 @@ PyTorch-based multi-task learning framework for discriminative modeling.
    LogWriter
    LogWriterConfig
    Operation
+   Scorer
    SnorkelClassifier
    Task
    TensorBoardWriter
