+++ snorkel/classification/__init__.py
@@ -1,6 +1,7 @@
 """PyTorch-based multi-task learning framework for discriminative modeling."""
 
 from .data import DictDataLoader, DictDataset  # noqa: F401
+from .scorer import Scorer  # noqa: F401
 from .snorkel_classifier import SnorkelClassifier  # noqa: F401
 from .task import Operation, Task, ce_loss, softmax  # noqa: F401
 from .training.loggers import (  # noqa: F401
